+++ packages/ui/src/lib/button/button.ts
@@ -2,8 +2,10 @@ import {
   booleanAttribute,
   ChangeDetectionStrategy,
   Component,
+  computed,
   input,
 } from '@angular/core';
+import { ElIcon, type ElIconSize, type ElIconVariant } from '../icon/icon';
 
 export type ElButtonVariant = 'primary' | 'secondary' | 'ghost';
 export type ElButtonSize = 'sm' | 'md' | 'lg';
@@ -11,6 +13,7 @@ export type ElButtonType = 'button' | 'submit' | 'reset';
 
 @Component({
   selector: 'el-button',
+  imports: [ElIcon],
   templateUrl: './button.html',
   styleUrl: './button.scss',
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -20,4 +23,9 @@ export class ElButton {
   readonly size = input<ElButtonSize>('md');
   readonly disabled = input(false, { transform: booleanAttribute });
   readonly type = input<ElButtonType>('button');
+  readonly iconStart = input('');
+  readonly iconEnd = input('');
+  readonly iconVariant = input<ElIconVariant>('solid');
+
+  protected readonly iconSize = computed((): ElIconSize => this.size());
 }